// src/common/enums/error-types.enum.ts
export enum ErrorTypes {
    BAD_REQUEST = 'Bad Request',
    
    INTERNAL_SERVER_ERROR = 'Internal Server Error',
    INVALID_CREDENTIALS = 'Invalid Credentials',
    RESOURCE_CONFLICT = 'Resource Conflict',
    RESOURCE_FORBIDDEN = 'Resource Forbidden',
    RESOURCE_NOT_FOUND = 'Resource Not Found',
    RESOURCE_NOT_VALID = 'Resource Not Valid',
    RESOURCE_UNAUTHORIZED = 'Resource Unauthorized',
}